mpris-widget: delete on mouse::enter code
this code is janky, because the notification is set asynchronously, it
can end up not being cleaned by the mouse exit and then you can end up
with multiple ones. Also, the image field of the table needs to be icon,
but if the icon is an http url like some players print or even a local
file like firefox does, setting the icon causes an error as it can't be
read.