battery-widget: Sanitize total charge computation
The code to parse the output of `acpi -i` is fragile as it does not
guarantee that the same number of battery status and capacities are
read. This causes a wrong computation of the total charge level when
batteries of status "Unknown" are reported.

For instance, this is what `acpi -i` reports with a Wacom Intuos S
attached to a Thinkpad T460s:

  $ acpi -i
  Battery 0: Unknown, 0%, rate information unavailable
  Battery 1: Not charging, 98%
  Battery 1: design capacity 1857 mAh, last full capacity 1513 mAh = 81%
  Battery 2: Not charging, 98%
  Battery 2: design capacity 2051 mAh, last full capacity 1711 mAh = 83%

Fix the code by guaranteeing that for each battery status parsed there
is also a capacity, which is zero if no capacity can be parsed. This
effectively causes the exclusion of such batteries when computing the
total charge level.