refactor: extract WeatherAPI widget into its own
This was done to allow for choice. In the first step I copied over the
existing code into a new folder and accompanied it with icons and
locales. There is no indicator of night in the JSON responses, which
made me delete the respective icons.

I copied and updated the README to fit the current implementation.
I will need to add screenshots before I can ask for another round of
review.

Signed-off-by: André Jaenisch <andre.jaenisch@posteo.de>