Use fahrenheit when units are imperial in weather-api-widget
I didn't see a good way to make a single check for units and set a
single temperature variable for each type (temp, feelslike, and forecast
min/max), since forecast use a loop to set multiple days in advance.
Instead, each location that displays a temperature uses an anonymous
function as a ternary operator to determine which to use.